+++ Minigin/minigin/utility/sprite.h
@@ -50,10 +50,10 @@ namespace mngn
         void set_left_offset_cols(int left);
         void set_top_offset_px(int top);
         void set_top_offset_rows(int top);
-        void set_collider_width(int collider_width);
-        void set_collider_height(int collider_height);
-        void set_collider_horizontal_offset(int collider_horizontal_offset);
-        void set_collider_vertical_offset(int collider_vertical_offset);
+        void set_collider_width(float collider_width);
+        void set_collider_height(float collider_height);
+        void set_collider_horizontal_offset(float collider_horizontal_offset);
+        void set_collider_vertical_offset(float collider_vertical_offset);
         void set_frames_per_sec(float frames_per_sec);
         void set_scale(float scale);
         void set_clip_width(int clip_width);
@@ -67,12 +67,12 @@ namespace mngn
         [[nodiscard]] auto texture() const -> texture *;
         [[nodiscard]] auto texture_width() const -> int;
         [[nodiscard]] auto texture_height() const -> int;
-        [[nodiscard]] auto collider_width() const -> int;
-        [[nodiscard]] auto collider_height() const -> int;
-        [[nodiscard]] auto collider_horizontal_offset() const -> int;
-        [[nodiscard]] auto collider_vertical_offset() const -> int;
-        [[nodiscard]] auto scaled_clip_width() const -> int;
-        [[nodiscard]] auto scaled_clip_height() const -> int;
+        [[nodiscard]] auto collider_width() const -> float;
+        [[nodiscard]] auto collider_height() const -> float;
+        [[nodiscard]] auto collider_horizontal_offset() const -> float;
+        [[nodiscard]] auto collider_vertical_offset() const -> float;
+        [[nodiscard]] auto scaled_clip_width() const -> float;
+        [[nodiscard]] auto scaled_clip_height() const -> float;
         [[nodiscard]] auto position() const -> glm::ivec2;
         [[nodiscard]] auto iter_count() const -> int;
         
@@ -106,10 +106,10 @@ namespace mngn
         int   clip_height_                = 0;
         int   grid_width_                 = 0;
         int   grid_height_                = 0;
-        int   collider_width_             = 0;
-        int   collider_height_            = 0;
-        int   collider_horizontal_offset_ = 0;
-        int   collider_vertical_offset_   = 0;
+        float collider_width_             = 0.0f;
+        float collider_height_            = 0.0f;
+        float collider_horizontal_offset_ = 0.0f;
+        float collider_vertical_offset_   = 0.0f;
         int   iter_count_                 = 0;
         
         bool    animated_         = false;
